--- a/modules/calib3d/doc/calib3d.bib
+++ b/modules/calib3d/doc/calib3d.bib
@@ -9,17 +9,6 @@
   publisher={Springer}
 }
 
-@article{gao2003complete,
-  title={Complete solution classification for the perspective-three-point problem},
-  author={Gao, Xiao-Shan and Hou, Xiao-Rong and Tang, Jianliang and Cheng, Hang-Fei},
-  journal={Pattern Analysis and Machine Intelligence, IEEE Transactions on},
-  volume={25},
-  number={8},
-  pages={930--943},
-  year={2003},
-  publisher={IEEE}
-}
-
 @inproceedings{hesch2011direct,
   title={A direct least-squares (DLS) method for PnP},
   author={Hesch, Joel and Roumeliotis, Stergios and others},
@@ -27,6 +16,17 @@
   pages={383--390},
   year={2011},
   organization={IEEE}
+}
+
+@misc{Pajdla2016,
+  author={Tomas Pajdla},
+  title={Elements of Geometry for Computer Vision},
+  year={2016},
+  month={May},
+  day={9},
+  pages={51--59},
+  publisher={Czech Technical University in Prague},
+  howpublished={\url{https://cmp.felk.cvut.cz/~pajdla/gvg/GVG-2016-Lecture.pdf}}
 }
 
 @article{penate2013exhaustive,
@@ -40,15 +40,6 @@
   publisher={IEEE}
 }
 
-@inproceedings{Terzakis2020SQPnP,
-  title={A Consistently Fast and Globally Optimal Solution to the Perspective-n-Point Problem},
-  author={George Terzakis and Manolis Lourakis},
-  booktitle={European Conference on Computer Vision},
-  pages={478--494},
-  year={2020},
-  publisher={Springer International Publishing}
-}
-
 @inproceedings{strobl2011iccv,
   title={More accurate pinhole camera calibration with imperfect planar target},
   author={Strobl, Klaus H. and Hirzinger, Gerd},
@@ -60,4 +51,13 @@
   publisher={IEEE},
   url={https://elib.dlr.de/71888/1/strobl_2011iccv.pdf},
   doi={10.1109/ICCVW.2011.6130369}
+}
+
+@inproceedings{Terzakis2020SQPnP,
+  title={A Consistently Fast and Globally Optimal Solution to the Perspective-n-Point Problem},
+  author={George Terzakis and Manolis Lourakis},
+  booktitle={European Conference on Computer Vision},
+  pages={478--494},
+  year={2020},
+  publisher={Springer International Publishing}
 }