… modules/calib3d/src/p3p.cpp