@s-trinh 2025-09-04
Looks like there are some accuracy issues with (my porting of) the USAC P3P.

---

Instead, I propose to use a more up-to-date P3P method, implemented here: https://github.com/PoseLib/PoseLib/blob/79fe59ada3122c50383ac06e043a5e04072c6711/PoseLib/solvers/p3p.cc

Results compared to the AP3P method and using a hand-held camera give good results (no visual difference, a little bit more computation time) compared to AP3P.

---

```bash
@inproceedings{ding2023revisiting,
  title={Revisiting the P3P Problem},
  author={Ding, Yaqing and Yang, Jian and Larsson, Viktor and Olsson, Carl and {\AA}str{\"o}m, Kalle},
  booktitle={Proceedings of the IEEE/CVF Conference on Computer Vision and Pattern Recognition},
  pages={4872--4880},
  year={2023}
}
```

Benchmark images from [here](https://github.com/yaqding/P3P):

<img width="1096" height="191" alt="image" src="https://github.com/user-attachments/assets/03e1d30e-9f7f-451f-8883-1565f10a5e35" />

<img width="739" height="169" alt="image" src="https://github.com/user-attachments/assets/7d37f05e-17c7-428f-8fbe-1b2a5e3c8576" />